Added buffer reduction algorithm in buffer::close (), found and fixed potencial bug in quantum_channel::make_noise()

+++ Classes/QC_classes/QC_classes/channels.cpp
@@ -18,10 +18,10 @@ void quantum_channel::make_noise(double max_noise)    // max_noise is a maximal 
 {
 	if (max_noise < 0 || max_noise > 100)
 		throw "Inappropriate max noise parameter";
-	double rand_iterator;
+	unsigned int rand_iterator;
 	bool temp;
 	for (int i = 0; i < static_cast<int>(max_noise*key_size/100); i++) {
-		rand_iterator = static_cast<int>((static_cast<double>(rand())/ RAND_MAX)*(key_size+1));
+		rand_iterator = static_cast<unsigned int>((static_cast<double>(rand())/ RAND_MAX)*(key_size+1));
 		if (rand_iterator == key_size)
 			rand_iterator--;
 		temp = buffer::randomize();
